requirements.txt
# Core Python  
python-dotenv==1.0.0  
loguru==0.7.0  
python-dateutil==2.8.2  
pytz==2023.3  

# Trading & Market Data  
ccxt==4.1.90  
ccxt-pro==0.9.77  
pandas==2.1.0  
numpy==1.26.0  
scipy==1.11.2  
statsmodels==0.14.0  
ta-lib==0.4.27  

# Machine Learning & NLP  
torch==2.1.0  
transformers==4.34.0  
sentencepiece==0.1.99  
protobuf==4.24.4  
scikit-learn==1.3.1  
onnxruntime==1.16.1  
huggingface-hub==0.17.3  

# APIs & Web Services  
aiohttp==3.8.6  
requests==2.31.0  
websockets==11.0.3  
uvicorn[standard]==0.23.2  
fastapi==0.103.1  
gunicorn==21.2.0  
python-socketio==5.9.0  

# Database & Storage  
sqlalchemy==2.0.21  
aiosqlite==0.19.0  
pyarrow==14.0.1  
redis==5.0.0  

# Configuration  
pyyaml==6.0.1  
tomli==2.0.1  
python-dotenv==1.0.0  

# Cryptography & Security  
cryptography==41.0.4  
pyopenssl==23.3.0  

# Monitoring & Metrics  
prometheus-client==0.17.1  
grafana-client==3.5.1  
sentry-sdk==1.34.0  

# Utilities  
tqdm==4.66.1  
joblib==1.3.2  
psutil==5.9.6  
sh==2.0.6  

# Backtesting  
backtrader==1.9.78.123  
quantconnect-stubs==0.1.1  

# Testing & Development  
pytest==7.4.2  
pytest-asyncio==0.21.1  
pytest-cov==4.1.0  
pytest-mock==3.11.1  
pytest-xdist==3.4.0  
faker==19.6.2  
responses==0.23.3  

# Formatting & Quality  
black==23.10.1  
flake8==6.1.0  
mypy==1.6.1  
isort==5.12.0  